backend/api/requirements.txt
django==1.9.6
psycopg2==2.6.1
djangorestframework==3.6.3
django-cors-middleware==1.3.1
django-filter==1.0.4
